change format to docx for now
the quarto -> docx process still has some hurdles when trying to use reference docx to match the desired style, etc. Will need a more thorough investigation in the future. But, for now, seems we should just go with HTML and keep the inertia rolling forward on creation of reproducible/functional qmd docs